Pull code related to video stream preferences into own file
This restructuring removes some duplicate code and is generally easier
to understand.
